Move RetryContext to core.retry package.
