+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/retry/Retryer.java
@@ -17,8 +17,6 @@
 
 package org.apache.eventmesh.runtime.core.retry;
 
-import org.apache.eventmesh.runtime.core.protocol.RetryContext;
-
 /**
  * Retryer interface
  */
